Reorganize storybook demo
