+++ .storybook/config.js
@@ -1,4 +1,5 @@
 import { configure } from '@storybook/react';
+
 function loadStories() {
   require('../stories/index.js');
   // You can require as many stories as you need.
